chore(posts): Move ItemLinks into helper. Fix TOC icon size.
TOC mobile icon now same as BackToTop.
