+++ src/components/posts/PostLayout.tsx
@@ -6,7 +6,7 @@ import Markdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 import rehypeRaw from 'rehype-raw';
 
-import ItemLinks from './ItemLinks';
+import ItemLinks from '../helpers/ItemLinks';
 
 import { getConfig } from '@/services/config';
